requirements.txt
influxdb<2.7.0
python-server-metrics
